QACI-451 Fixed jaxws-tracing compilation and deployment issues
- fixed crash on JDK11
- fixed weird exception on JDK8 related to Unstable
- separated test and deployment module packages as they use different
  compiler output paths
