+++ appserver/tests/payara-samples/samples/jaxws-tracing/src/main/java/fish/payara/samples/jaxws/endpoint/servlet/JAXWSEndPointImplementation.java
@@ -39,6 +39,9 @@
  */
 package fish.payara.samples.jaxws.endpoint.servlet;
 
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
 import javax.jws.WebService;
 
 import org.eclipse.microprofile.opentracing.Traced;
@@ -47,10 +50,12 @@ import org.eclipse.microprofile.opentracing.Traced;
     endpointInterface = "fish.payara.samples.jaxws.endpoint.servlet.JAXWSEndPointInterface",
     serviceName = "JAXWSEndPointImplementationService")
 public class JAXWSEndPointImplementation implements JAXWSEndPointInterface {
+    private static final Logger LOG = Logger.getLogger(JAXWSEndPointImplementation.class.getName());
 
     @Override
     @Traced(operationName = "customOperation")
     public String sayHi(String name) {
+        LOG.log(Level.INFO, "sayHi(name={0})", name);
         return "Hi " + name;
     }
